Use mode 'ni' for nvim_feedkeys
When mode is 'n' then the keys are _appended_ to the typeahead buffer,
which can add them _after_ other keys in a mapping. Using 'ni' instead
inserts them, so that keys later in the binding remain later in the
buffer.

This allows mappings such as:

"gcATODO<ESC>"

to work as expected.