fix: respect updated `v:count` when using `.` repeat (#253)
Previously, when you press `3gcc` 3 lines will be (un)commented then pressing `.` also (un)comments 3 lines. This is fine. But if you press `4.` (prefixing the `.` with a `count`) it (un)comments 3 lines instead of 4 which is wrong. As `.` repeat should respect the updated value of the `v:count`, in this case, `4` should be used instead of `3`.

With this PR, `.` will respect the updated value of `v:count` by removing an old workaround. This PR also replaces `vim.v.count` with `vim.api.nvim_get_vvar('count')`.